Refactored models.py into a package.
