+++ run/models/__init__.py
@@ -1,0 +1,12 @@
+"""Initialise the models package."""
+
+from .poetry_models import Meter, Poet, Poem
+from .user_models import Role, User
+
+__all__ = [
+    "Meter",
+    "Poet",
+    "Poem",
+    "Role",
+    "User",
+]
